% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/area.diff.R
\name{area.diff}
\alias{area.diff}
\title{Area difference}
\usage{
area.diff(x, y, rarefy, cent.tend = mean)
}
\arguments{
\item{x, y}{the two distributions to compare.}

\item{rarefy}{Optional, if \code{x != y}, how many rarefaction samples to use. If left empty a default number of replicates between 100 and 1000 is used (see details).}

\item{cent.tend}{Optional, if \code{x != y}, which central tendency to use (\code{defaut = mean})}
}
\description{
Measure the area difference between two ranked distribution
}
\details{
The number of replicates is chosen based on the variance of the distribution to rarefy using the Silverman's rule of thumb (Silverman 1986, pp.48, eqn 3.31) for choosing the bandwidth of a Gaussian kernel density estimator multiplied by 1000 with a result comprised between 100 and 1000.
For example, for rarefying \code{x}, \code{rarefy = round(stats::bw.nrd0(x) * 1000)}.
With \code{100 <= rarefy <= 1000}.
}
\examples{
## Two distributions
x <- rnorm(10)
y <- runif(10)

## The area difference
area.diff(x, y)

}
\references{
Silverman, B. W. (1986) Density Estimation. London: Chapman and Hall.
}
\seealso{
\code{\link{variation.range}}
}
\author{
Thomas Guillerme
}
