% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/landmarktest-package.R
\docType{package}
\name{landmarktest-package}
\alias{landmarktest-package}
\title{Wrapping Package For Performing The Landmark Test From Papername}
\description{
This wrapping package contains the functions for performing the landmark tests from @@ PAPER.
}
\author{
Thomas Guillerme <guillert@tcd.ie>
}
\keyword{disparity,}
\keyword{geometric}
\keyword{landmark,}
\keyword{morphometric,}
\keyword{permutation}
\keyword{random}
\keyword{test}
\keyword{test,}
